Standardize BND File location to typical Gradle src/ and build/ dirs
This moves the template file to a src/ subfolder, as is the Gradle
standard for input files that are processed.

This also moves the output file to build/, the Gradle standard for
processed files that tasks produce.

Additionally, tell the Bundle logic where to find the relocated bnd file
after it has been generated, and fix the dependsOn ordering because
prepareBndFile actually does depend on copyDependencies. Remove bnd.bnd
from gitignore, because it's no longer necessary.