Complete the migration of WSI-DI to CSV-based IO
Previous versions of the WSI-DI method relied on libraries that
exclusively worked with DSS Version 6. As a part of the feature update
efforts for WRIMS to support DSS 7, the WSI-DI scripts were updated to
work  without these libraries. The solution was to perform all IO without
DSS interations, instead relying on CSV files. WRIMS already had the
feature to output DV files to CSV rather than DSS, so the new method
only requires `.launch` configurations to target CSV, rather than DSS.

Since the updated method only changed the parts of the WSI-DI method
related to IO, other changes (including code cleanup) were not made at
this time. Some non-essential changes were made (primarily to
spaces/tabs inconsistencies), but they should not impact the WSI-DI
process.

The reason for not performing a larger code update is that the WSI-DI
method is on it's way to being depreciated in favor of alternative
methods, namely the "CalSim Allocation Module (CAM)".